requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.1.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
tomli==2.2.1
tzdata==2025.2
urllib3==2.4.0
